Code gen

+++ include/Actions.h
@@ -59,7 +59,7 @@ class Actions {
     Operator, std::shared_ptr<Expr>, std::shared_ptr<Expr>
   );
   std::shared_ptr<Expr> unary(symbol::Tag, std::shared_ptr<Expr>);
-  std::shared_ptr<Expr> constant(symbol::Tag t, int val = 0, double dec = 0.0);
+  std::shared_ptr<Expr> constant(symbol::Tag t, int val = 0, int dec = 0);
   std::shared_ptr<Expr> empty_expr() { return std::make_shared<Expr>(Type()); }
 
   // helpers
